move the integration tests into separate sourceSets
